Add plugin system.
This is more elaborate than the first prototype. Many plugins can
coexist in the plugins directory. Each one is a module which can
contain three functions. Any of these functions can be safely omitted
and it will do nothing.

The init function is called when gotosocial starts up; it takes no args.

The process function is called when a Status is saved to the DB. Its
only argument is the status itself. It doesn't return anything, but it
can make changes to the status fields.

Finally the run function can be triggered from the command line:

    $ ./gotosocial plugin foo arg1 arg2 arg3

This will trigger the run function of the foo function; it will be
called with an array of strings that in the above example would be the
three args. The string array is a userdata on the Lua side, but it can
be indexed using args[n] or called as a function to produce an iterator.

At this point the main shortcoming is that the actions which can be
taken by a plugin are limited to just printing and making changes to
the statuses, which makes this of limited utility. However, the next
step could be to expose gotosocial functionality to the Lua side such
as following and unfollowing, blocking, or even posting replies.

Here is an example plugin:

local lols = {}

local function lol(uri)
   local count = lols[uri]

   if count then
      lols[uri] = count + 1
   else
      lols[uri] = 1
   end
end

local function unlol(uri)
   local count = lols[uri]

   if count then
      print(string.format("%s-lol chain! Marvellous!", count))
      lols[uri] = nil
   end
end

return {
   run=function(args)
      for _,x in args() do print("ran with", x) end
   end,

   process = function(status)
      if(status.Content:find("lol")) then
         lol(status.AccountURI)
      else
         unlol(status.AccountURI)
      end
   end,

   init = function()
      print("Plugin lol initialized!")
   end,
}